requirements.txt
gradio
opencv-python
numpy
Pillow
